+++ Model/ReglasNegocio.cs
@@ -6,7 +6,7 @@ namespace Model
 {
     public class ReglasNegocio
     {
-        public long ReglasNegocioId { get; set; }
+        public int ReglasNegocioId { get; set; }
         public DateTime FechaActualizacion { get; set; }
         public float Tarifa { get; set; }
         public float PrecioKM { get; set; }
